Re-committed my branch, leaving just my modified KindInput in a new folder to avoid conflicts
